feat: implement auth-jwt

+++ courses-app-ts/api/src/controllers/courseController/CourseController.ts
@@ -27,6 +27,9 @@ class CourseController {
 
     public async addComment(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { userId, content, courseId }: ICreateCommentDto = req.body;
             if (!userId || !content || !courseId) {
                 throw new AppError('UserId, content, and courseId are required', 400);
@@ -49,6 +52,9 @@ class CourseController {
 
     public async getComments(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { courseId } = req.params;
             const comments = await this.commentService.findCommentsByCourseId(courseId as string);
             res.status(200).send(comments);
@@ -59,6 +65,9 @@ class CourseController {
 
     public async deleteComment(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { commentId } = req.params;
             const deletedComment = await this.commentService.deleteComment(commentId);
             res.status(200).send({ message: 'Comment deleted successfully', deletedComment });
@@ -69,6 +78,9 @@ class CourseController {
 
     public async addRating(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { userId, rating, courseId }: ICreateRatingDto = req.body;
             if (!userId || !rating || !courseId) {
                 throw new AppError('UserId, rating, and courseId are required', 400);
@@ -92,6 +104,9 @@ class CourseController {
 
     public async getRatings(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { courseId } = req.params;
             if (!courseId) {
                 throw new AppError('CourseId is required', 400);
@@ -105,6 +120,9 @@ class CourseController {
 
     public async deleteRating(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { ratingId } = req.params;
             const deletedRating = await this.ratingService.deleteRating(ratingId);
             if (!deletedRating) {
@@ -118,8 +136,11 @@ class CourseController {
 
     public async getAllCourses(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const courses = await this.coursesService.findCourses();
-            res.status(200).send(courses);
+            res.status(200).send({ courses });
         } catch (err) {
             next(err);
         }
@@ -127,6 +148,9 @@ class CourseController {
 
     public async getCourseById(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { courseId } = req.params;
             const courses = await this.coursesService.findCourseById(courseId);
             if (!courses) {
@@ -140,6 +164,9 @@ class CourseController {
 
     public async deleteCourse(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { courseId } = req.params;
             const deletedCourse = await this.coursesService.deleteCourseById(courseId);
             if (!deletedCourse) {
@@ -153,6 +180,9 @@ class CourseController {
 
     public async createCourse(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { title, description, difficulty }: ICreateCourseDto = req.body;
             const courseId = await this.coursesService.createCourse({
                 title,
@@ -171,6 +201,9 @@ class CourseController {
 
     public async updateCourse(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
+            if (!req.user) {
+                throw new AppError('Not authorized', 401);
+            }
             const { courseId } = req.params;
             const updateData: IUpdateCourseDTO = req.body;
             const updateCourse = await this.coursesService.updateCourseById(courseId, updateData);
